+++ src/nodePluginSdk.ts
@@ -2,3 +2,5 @@ export * from "./nodeSpecSdk";
 export * from "./nodePluginPackages";
 
 export * from "./nodePluginArchive";
+
+export * from "./nodePluginResources";